Combine node implements into one source file, move helpers
