+++ pkg/config/nodetreemodel/helpers.go
@@ -9,8 +9,11 @@ import (
 	"fmt"
 	"reflect"
 	"strings"
+	"time"
 	"unicode"
 
+	"github.com/DataDog/datadog-agent/pkg/config/helper"
+	"github.com/DataDog/datadog-agent/pkg/config/model"
 	"github.com/spf13/cast"
 )
 
@@ -93,4 +96,89 @@ func ToMapStringInterface(data any, path string) (map[string]interface{}, error)
 		return convert, nil
 	}
 	return nil, fmt.Errorf("expected map at '%s' got: %v", path, v)
+}
+
+func mapToMapString(m reflect.Value) map[string]interface{} {
+	if v, ok := m.Interface().(map[string]interface{}); ok {
+		// no need to convert the map
+		return v
+	}
+
+	res := make(map[string]interface{}, m.Len())
+
+	iter := m.MapRange()
+	for iter.Next() {
+		k := iter.Key()
+		mk := ""
+		if k.Kind() == reflect.String {
+			mk = k.Interface().(string)
+		} else {
+			mk = fmt.Sprintf("%s", k.Interface())
+		}
+		res[mk] = iter.Value().Interface()
+	}
+	return res
+}
+
+// NewNodeTree will recursively create nodes from the input value to construct a tree
+func NewNodeTree(v interface{}, source model.Source) (*nodeImpl, error) {
+	if helper.IsNilValue(v) {
+		// nil as a value acts as the zero value, and the cast library will correctly
+		// convert it to zero values for the types we handle
+		return newLeafNode(nil, source), nil
+	}
+	switch it := v.(type) {
+	case []interface{}:
+		return newLeafNode(it, source), nil
+	}
+
+	// handle all map types that can be converted to map[string]interface{}
+	if v := reflect.ValueOf(v); v.Kind() == reflect.Map {
+		children, err := makeChildNodeTrees(mapToMapString(v), source)
+		if err != nil {
+			return nil, err
+		}
+		return newInnerNode(children), nil
+	}
+
+	if isScalar(v) {
+		return newLeafNode(v, source), nil
+	}
+	// Finally, try determining node type using reflection, should only be needed for unit tests that
+	// supply data that isn't one of the "plain" types produced by parsing json, yaml, etc
+	node, err := asReflectionNode(v)
+	if err == errUnknownConversion {
+		return nil, fmt.Errorf("could not create node from: %v of type %T", v, v)
+	}
+	return node, err
+}
+
+func makeChildNodeTrees(input map[string]interface{}, source model.Source) (map[string]*nodeImpl, error) {
+	children := make(map[string]*nodeImpl)
+	for k, v := range input {
+		node, err := NewNodeTree(v, source)
+		if err != nil {
+			return nil, err
+		}
+		children[k] = node
+	}
+	return children, nil
+}
+
+func isScalar(v interface{}) bool {
+	switch v.(type) {
+	case int, int8, int16, int32, int64:
+		return true
+	case uint, uint8, uint16, uint32, uint64:
+		return true
+	case bool, string, float32, float64, time.Time, time.Duration:
+		return true
+	default:
+		return false
+	}
+}
+
+func isSlice(v interface{}) bool {
+	rval := reflect.ValueOf(v)
+	return rval.Kind() == reflect.Slice
 }
